# backend/src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: ちいぴよ API
  description: 子供の写真・動画管理及び育児記録管理アプリケーションのAPI定義
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /media:
    get:
      tags:
        - mediaManagement
      summary: メディア一覧を取得
      operationId: getMediaList
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/MediaTypeFilter'
        - $ref: '#/components/parameters/AlbumIdFilter'
        - $ref: '#/components/parameters/AlbumIdExclude'
        - $ref: '#/components/parameters/TagIdFilter'
        - $ref: '#/components/parameters/SharingGroupIdFilter'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/FavoriteFilter'
      responses:
        '200':
          description: メディア一覧
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/MediaListResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - mediaManagement
      summary: メディアのメタデータを登録し署名付きアップロードURLを取得
      operationId: createMedia
      x-codegen-request-body-name: MediaUploadData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: アップロードするメディアのメタデータ
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaUploadRequestDto'
        required: true
      responses:
        '201':
          description: メタデータ登録成功、署名付きURLを返却
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/MediaUploadResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - mediaManagement
      summary: 複数メディアを削除
      description: ゴミ箱に移動し30日間保持
      operationId: deleteMultipleMedia
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/MediaIds'
      responses:
        '204':
          description: メディアの削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /media/{id}:
    get:
      tags:
        - mediaManagement
      summary: メディアをID指定で1件取得
      operationId: getMedia
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: メディア情報
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    patch:
      tags:
        - mediaManagement
      summary: メディア情報を更新
      operationId: updateMedia
      x-codegen-request-body-name: MediaUpdateData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 更新するメディア情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaUpdateRequestDto'
        required: true
      responses:
        '204':
          description: メディアの更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - mediaManagement
      summary: メディアを削除
      description: ゴミ箱に移動し30日間保持
      operationId: deleteMedia
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: メディアの削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /media/{id}/status:
    patch:
      tags:
        - mediaManagement
      summary: メディアのアップロード状態を更新
      operationId: updateMediaUploadStatus
      x-codegen-request-body-name: MediaUpdateStatusData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 更新するステータス情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaUploadStatusRequestDto'
        required: true
      responses:
        '200':
          description: 更新されたメディア情報
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/MediaResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /media/batch:
    patch:
      tags:
        - mediaManagement
      summary: メディアの一括更新
      operationId: updateMediaBatch
      x-codegen-request-body-name: MediaBatchUpdateData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: 一括更新するメディアのメタデータ
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/MediaBatchUpdateRequestDto'
        required: true
      responses:
        '204':
          description: 複数メディアの更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognitoで取得したJWTトークンを指定
  parameters:
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        default: XMLHttpRequest
    ResourceId:
      name: id
      in: path
      description: リソースの一意な識別子
      required: true
      schema:
        type: integer
        format: int64
    PageOffset:
      name: offset
      in: query
      description: 取得開始位置
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    PageLimit:
      name: limit
      in: query
      description: 取得件数
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    MediaTypeFilter:
      name: mediaKind
      in: query
      description: メディア種別で絞り込み
      required: false
      schema:
        type: string
        enum:
          - PHOTO
          - VIDEO
    AlbumIdFilter:
      name: albumId
      in: query
      description: アルバムIDで絞り込み
      required: false
      schema:
        type: integer
        format: int64
    AlbumIdExclude:
      name: excludeAlbumId
      in: query
      description: アルバムIDで除外
      required: false
      schema:
        type: integer
        format: int64
    TagIdFilter:
      name: tagId
      in: query
      description: タグIDで絞り込み（複数指定可、OR検索）
      required: false
      schema:
        type: array
        items:
          type: integer
          format: int64
    SharingGroupIdFilter:
      name: sharingGroupId
      in: query
      description: 共有範囲グループIDで絞り込み
      required: false
      schema:
        type: integer
        format: int64
    StartDate:
      name: startDate
      in: query
      description: 検索開始日
      required: false
      schema:
        type: string
        format: date
    EndDate:
      name: endDate
      in: query
      description: 検索終了日
      required: false
      schema:
        type: string
        format: date
    FavoriteFilter:
      name: isFavorite
      in: query
      description: お気に入りフィルタ
      required: false
      schema:
        type: boolean
    MediaIds:
      name: mediaIds
      in: query
      description: メディアのIDリスト
      required: true
      schema:
        type: array
        items:
          type: integer
          format: int64
...
  schemas:
    ErrorResponseDto:
      type: object
      description: エラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストが不正です
...
    TagResponseDto:
      type: object
      description: タグ情報
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: タグID
          example: 1
        name:
          type: string
          description: タグ名
          example: お出かけ
        mediaCount:
          type: integer
          format: int64
          description: タグが紐づいているメディアの数
          example: 10
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
    MediaNavigationResponseDto:
      type: object
      description: 前後のメディア情報
      required:
        - id
        - mediaType
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3アップロード用の署名付きURL
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
    MediaResponseDto:
      type: object
      description: メディア情報
      required:
        - id
        - uploadedBy
        - mediaType
        - originalFilename
        - contentType
        - fileSize
        - width
        - height
        - takenAt
        - albumId
        - sharingGroupId
        - uploadStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: メディアID
          example: 1
        uploadedBy:
          type: integer
          format: int64
          description: アップロードしたユーザーID
          example: 1
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
        originalFilename:
          type: string
          description: ファイル名
          example: IMG_20260101.jpg
        contentType:
          type: string
          description: MIMEタイプ
          example: image/jpeg
        fileSize:
          type: integer
          format: int64
          description: ファイルサイズ（バイト）
          example: 2048000
        width:
          type:
            - integer
            - 'null'
          description: 幅（ピクセル）
          example: 1920
        height:
          type:
            - integer
            - 'null'
          description: 高さ（ピクセル）
          example: 1080
        presignedUrl:
          type: string
          format: uri
          description: S3署名付きURL
        thumbnailPresignedUrl:
          type:
            - string
            - 'null'
          format: uri
          description: サムネイルのS3署名付きURL
        takenAt:
          type:
            - string
            - 'null'
          format: date
          description: 撮影日
          example: '2026-01-01'
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
          example: 1
        sharingGroupId:
          type: integer
          format: int64
          description: 共有範囲グループID
          example: 1
        uploadStatus:
          type: string
          description: アップロード状況
          enum:
            - PROCESSING
            - COMPLETED
            - FAILED
          example: COMPLETED
        addFavoriteUserIds:
          type: array
          description: お気に入りに追加したユーザーID一覧
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
        tags:
          type: array
          description: 付与されているタグ一覧
          items:
            $ref: '#/components/schemas/TagResponseDto'
        isFavorite:
          type: boolean
          description: お気に入り済みフラグ
          example: false
        commentCount:
          type: integer
          format: int64
          description: コメント数
          example: 1
        nextMedia:
          type:
            - object
            - 'null'
          description: 1つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondNextMedia:
          type:
            - object
            - 'null'
          description: 2つ後のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        previousMedia:
          type:
            - object
            - 'null'
          description: 1つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        secondPreviousMedia:
          type:
            - object
            - 'null'
          description: 2つ前のメディア
          $ref: '#/components/schemas/MediaNavigationResponseDto'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
    MediaListResponseDto:
      type: object
      description: メディア一覧レスポンス
      required:
        - items
        - totalCount
        - hasNext
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MediaResponseDto'
        totalCount:
          type: integer
          format: int64
          description: 総件数
          example: 150
        hasNext:
          type: boolean
          description: 次のページが存在するかどうか
          example: true
    MediaUploadRequestDto:
      type: object
      description: メディアアップロードリクエスト
      required:
        - mediaType
        - originalFilename
        - contentType
        - fileSize
      properties:
        mediaType:
          type: string
          description: メディア種別
          enum:
            - PHOTO
            - VIDEO
          example: PHOTO
        originalFilename:
          type: string
          description: ファイル名
          maxLength: 255
          example: IMG_20260101.jpg
        contentType:
          type: string
          description: MIMEタイプ
          maxLength: 100
          example: image/jpeg
        fileSize:
          type: integer
          format: int64
          description: ファイルサイズ（バイト）
          example: 2048000
        width:
          type:
            - integer
            - 'null'
          description: 幅（ピクセル）
        height:
          type:
            - integer
            - 'null'
          description: 高さ（ピクセル）
        takenAt:
          type:
            - string
            - 'null'
          format: date
          description: 撮影日
          example: '2026-01-01'
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
        sharingGroupId:
          type:
            - integer
            - 'null'
          format: int64
          description: 共有範囲グループID
          example: 1
    MediaUploadResponseDto:
      type: object
      description: メディアアップロードレスポンス
      required:
        - mediaId
        - presignedUrl
      properties:
        mediaId:
          type: integer
          format: int64
          description: 作成されたメディアID
          example: 1
        presignedUrl:
          type: string
          format: uri
          description: S3アップロード用の署名付きURL
          example: https://s3.amazonaws.com/bucket/...
    MediaUpdateRequestDto:
      type: object
      description: メディア更新リクエスト
      properties:
        albumId:
          type:
            - integer
            - 'null'
          format: int64
          description: 所属アルバムID
        sharingGroupId:
          type:
            - integer
            - 'null'
          format: int64
          description: 共有範囲グループID
    MediaUploadStatusRequestDto:
      type: object
      description: メディアアップロードステータス更新リクエスト
      required:
        - uploadStatus
      properties:
        uploadStatus:
          type: string
          description: アップロード状況
          enum:
            - PROCESSING
            - COMPLETED
            - FAILED
          example: COMPLETED
    MediaBatchUpdateRequestDto:
      type: object
      description: メディア一括更新リクエスト
      required:
        - mediaIds
      properties:
        mediaIds:
          type: array
          description: メディアIDの配列
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
        sharingGroupId:
          type:
            - integer
            - 'null'
          format: int64
          description: 共有範囲グループID
        tagIds:
          type:
            - array
            - 'null'
          description: タグIDの配列
          items:
            type: integer
            format: int64
          example:
            - 1
            - 2
            - 3
...
  responses:
    DefaultErrors:
      description: 予期しないエラーが発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Unauthorized:
      description: 認証エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    NotFound:
      description: 指定されたリソースが見つからない
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    InternalServerError:
      description: サーバー内部でエラー発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Forbidden:
      description: アクセス権限エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'